Bug #1130501 bsp: wcn: Fix problem when restore p2p mac addr
[root cause  ] zero mac addr did not processed
[changes     ] wifi driver
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:s9863a1h10
[test case]:vts
[reviewers   ] keguang.zhang, baolei.yuan
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

The vts case SupplicantP2pIfaceHidlTest.EnableMacRandomization
may set P2P addr to zero, which means restore the addr to original.
the problem is that function is_valid_ether_addr estimate the zero
addr as a invalid value, and do nothing. So we use is_zero_ether_addr
to estimate whether the addr is zero, if the addr is zero, we set the
addr to the original vif addr.

Change-Id: I98a0f379c3a2b9072ab867a2aca4f0d8b9dcea53
Signed-off-by: Gui Zhu <gui.zhu@spreadtrum.com>